getrequiredepartment
